Split MountPoint into platform specific
